pyproject.toml
# SPDX-License-Identifier: BSD-3-Clause
# Copyright Contributors to the OpenColorIO Project.

[tool.poetry]
name = "opencolorio-config-aces"
version = "2.0.0"
description = "OpenColorIO Configuration for ACES"
authors = ["OpenColorIO Contributors"]
maintainers = ["OpenColorIO Contributors <ocio-dev@lists.aswf.io>"]
license = "BSD-3-Clause"
readme = 'README.rst'
repository = "https://github.com/AcademySoftwareFoundation/OpenColorIO-Config-ACES"
homepage = "https://www.colour-science.org/"
keywords = [
    "ACES",
    "OpenColorIO",
    "OCIO"
]

[tool.poetry.dependencies]
python = ">= 3.9, < 3.12"
opencolorio = ">= 2, < 3"
requests = ">= 2, < 3"
semver = ">= 3, < 4"

[tool.poetry.group.optional.dependencies]
colour-science = ">= 0.4.3"
jsonpickle = ">= 2, < 3"
networkx = "*"

[tool.poetry.group.graphviz.dependencies]
pygraphviz = "*"

[tool.poetry.group.dev.dependencies]
coverage = "*"
coveralls = "*"
invoke = "*"
pre-commit = "*"
pydata-sphinx-theme = "*"
pyright = "*"
pytest = "*"
pytest-cov = "*"
restructuredtext-lint = "*"
sphinx = ">= 4, < 5"
twine = "*"

[tool.poetry.group.docs.dependencies]
pydata-sphinx-theme = "*"
sphinx = ">= 4, < 5"

[tool.flynt]
line_length = 999

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 88
multi_line_output = 3
skip = ["__init__.py"]
split_on_trailing_comma = true
use_parentheses = true

[tool.pyright]
reportMissingImports = false
reportMissingModuleSource = false
reportUnboundVariable = false
reportUnnecessaryCast = true
reportUnnecessaryTypeIgnoreComment = true
reportUnsupportedDunderAll = false
reportUnusedExpression = false
exclude = ["**/aces-dev"]

[tool.ruff]
target-version = "py39"
line-length = 88
select = [
    "A", # flake8-builtins
    "ARG", # flake8-unused-arguments
    # "ANN",  # flake8-annotations
    "B", # flake8-bugbear
    # "BLE",  # flake8-blind-except
    "C4", # flake8-comprehensions
    # "C90",  # mccabe
    # "COM",  # flake8-commas
    "DTZ", # flake8-datetimez
    "D", # pydocstyle
    "E", # pydocstyle
    # "ERA",  # eradicate
    # "EM",   # flake8-errmsg
    "EXE", # flake8-executable
    "F", # flake8
    # "FBT",  # flake8-boolean-trap
    "G", # flake8-logging-format
    "I", # isort
    "ICN", # flake8-import-conventions
    "INP", # flake8-no-pep420
    "ISC", # flake8-implicit-str-concat
    "N", # pep8-naming
    # "PD",   # pandas-vet
    "PIE", # flake8-pie
    "PGH", # pygrep-hooks
    "PL", # pylint
    # "PT",   # flake8-pytest-style
    # "PTH",  # flake8-use-pathlib          [Enable]
    "Q", # flake8-quotes
    "RET", # flake8-return
    "RUF", # Ruff
    "S", # flake8-bandit
    "SIM", # flake8-simplify
    "T10", # flake8-debugger
    "T20", # flake8-print
    # "TCH",  # flake8-type-checking
    "TID", # flake8-tidy-imports
    "TRY", # tryceratops
    "UP", # pyupgrade
    "W", # pydocstyle
    "YTT"     # flake8-2020
]
ignore = [
    "B008",
    "B905",
    "D104",
    "D200",
    "D202",
    "D205",
    "D301",
    "D400",
    "I001",
    "N801",
    "N802",
    "N803",
    "N806",
    "N813",
    "N815",
    "N816",
    "PGH003",
    "PIE804",
    "PLE0605",
    "PLR0912",
    "PLR0913",
    "PLR0915",
    "PLR2004",
    "PLW2901",
    "RET504",
    "RET505",
    "RET506",
    "RET507",
    "RET508",
    "TRY003",
    "TRY300",
]
fixable = ["B", "C", "E", "F", "PIE", "RUF", "SIM", "UP", "W"]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.per-file-ignores]
"docs/*" = ["INP"]
"tasks.py" = ["INP"]
"utilities/*" = ["EXE001", "INP"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
